# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Holodeck Router UI API",
    "description": "REST API for managing VCF 9 Holodeck deployments, Day 2 operations, reservations, commands, user management, and email notifications.",
    "version": "1.2.0"
  },
  "servers": [
    { "url": "/", "description": "Current instance" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "holodeck-token",
        "description": "JWT token set as httpOnly cookie after login"
      }
    },
    "schemas": {
...
      "Reservation": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "notes": { "type": "string", "nullable": true },
          "isMaintenance": { "type": "boolean" },
          "isCustomerDemo": { "type": "boolean" },
          "status": { "type": "string" },
          "userId": { "type": "string" },
          "user": {
            "type": "object",
            "properties": {
              "displayName": { "type": "string" },
              "username": { "type": "string" },
              "email": { "type": "string", "nullable": true }
            }
          }
        }
      },
...
    }
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/reservations": {
      "get": {
        "tags": ["Reservations"],
        "summary": "List reservations",
        "parameters": [
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "Filter: endTime >= from" },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "Filter: startTime <= to" }
        ],
        "responses": {
          "200": {
            "description": "Reservation list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reservations": { "type": "array", "items": { "$ref": "#/components/schemas/Reservation" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Reservations"],
        "summary": "Create a reservation",
        "description": "Creates a time slot. Returns overlap warning (409) if conflicts exist — resend with confirmOverlap: true to proceed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "startTime", "endTime"],
                "properties": {
                  "title": { "type": "string", "maxLength": 200 },
                  "startTime": { "type": "string", "format": "date-time" },
                  "endTime": { "type": "string", "format": "date-time" },
                  "notes": { "type": "string", "maxLength": 1000 },
                  "isMaintenance": { "type": "boolean", "description": "labadmin+ only" },
                  "isCustomerDemo": { "type": "boolean" },
                  "confirmOverlap": { "type": "boolean", "description": "Set true to bypass overlap warning" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Reservation created" },
          "409": { "description": "Overlap detected" }
        }
      }
    },
...
  }
}